Combine and restructure all vim configs into one module
Lunarvim's neovim settings were put into a proper nvim/init.lua, which
gets sourced in lvim's config.lua. The nvim/init.lua file sources the
original .vimrc for all the settings.

The plugin section in the .vimrc is gated behind (the non-existence of)
a flag to not install them when sourcing from certain locations. The
plugins in the .vimrc *are* installed when sourced from the
nvim/init.lua but *are not* when sourcing from the lvim/config.lua
file. This way I get the vim plugins for a normal neovim instance but
not when opening Lunarvim.