kwin_scripts: make firefox-pip-tiling work with flatpak
For some reason the flatpak version marks its pip windows as not
minimizable. This is a readonly property, so tough luck.

Workaround is to move the windows far off screen, then return them to
their original location. I don't get neat minimization/restoration
animations this way---the window just disappears and reappears---but it
works at least.